Apply random y velocity to asteroids, fix center
Asteroids were centered using the outer panel, but due to the padding
on this panel, the position relative to the inner panel was incorrect.

Random y velocity prevents asteroids from all bunching up in the center.
The 3/8 constant is entirely arbitrary, and simply chosen by comparing
in-game behavior to this task's behavior, and adjusting accordingly
until things "felt right."